+++ src/services/root-reducer.ts
@@ -2,7 +2,7 @@ import { combineReducers } from "@reduxjs/toolkit";
 
 import burgerDetails from "./reducers/ingredient-details";
 import allIngredients from "./reducers/burger-ingredients";
-import constructorIngredients from "./reducers/burger-constructor";
+import constructorIngredients from "./reducers/burger-constructor/burger-constructor";
 import orderDetails from "./reducers/order-details";
 import authUser from "./reducers/auth/auth";
 import { webSocketApi } from "./rtk/web-socket";
